% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_ignore_row.R
\name{cbind_ignore_row}
\alias{cbind_ignore_row}
\title{Combine columns of tibbles ignoring column length}
\usage{
cbind_ignore_row(df1, df2)
}
\arguments{
\item{df1}{A data frame or tibble.}

\item{df2}{A data frame or tibble.}
}
\value{
A data frame or tibble.
}
\description{
Equates the number of rows between two tibbles (by inserting NA rows)
and returns the column-combined result.
}
