% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_cols.R
\name{drop_na_cols}
\alias{drop_na_cols}
\title{Drop NA columns}
\usage{
drop_na_cols(df, percent_incomplete = 100)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{percent_incomplete}{A numeric value.}
}
\value{
A data frame or tibble.
}
\description{
Returns a tibble or data frame without columns containing only NA values. If
`percent_incomplete` is supplied, any column which has a total percentage of NA
values greater than or equal to `percent_incomplete` will be dropped.
}
\examples{
library(tidyr)

sample_data <-
  tibble(
    head(iris, 10),
    NA_100p = rep(NA, 10),
    NA_90p = c(1, rep(NA, 9))
  )

sample_data

drop_na_cols(sample_data)
drop_na_cols(sample_data, percent_incomplete = 90)
}
\seealso{
[drop_na_rows()]
}
