% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_rows.R
\name{drop_na_rows}
\alias{drop_na_rows}
\title{Drop NA rows}
\usage{
drop_na_rows(df, percent_incomplete = 100)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{percent_incomplete}{A numeric value.}
}
\value{
A data frame or tibble.
}
\description{
Returns a tibble or data frame without rows containing only NA values. If
`percent_incomplete` is supplied, any row which has a total percentage of NA
values greater than or equal to `percent_incomplete` will be dropped.
}
\examples{
library(tidyr)

sample_data <-
  tibble(rbind(
    head(iris, 5),
    rep(NA, 5),
    c(1, rep(NA, 4)))
  )

sample_data

drop_na_rows(sample_data)
drop_na_rows(sample_data, percent_incomplete = 75)
}
\seealso{
[drop_na_cols()]
}
