% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_matching.R
\name{files_matching}
\alias{files_matching}
\title{Locate files in a directory}
\usage{
files_matching(path, file_type, name_contains = NULL, ignore.case = TRUE)
}
\arguments{
\item{path}{The path to the directory to locate files.}

\item{file_type}{A character vector indicating a file extension.}

\item{name_contains}{An optional character vector to filter for matching file names.}

\item{ignore.case}{If \code{TRUE}, \code{name_contains} case is ignored when filtering matching file names.}
}
\value{
A three column tibble.
}
\description{
Search a directory for a given file type and name, return
a tibble of matching file information sorted by modified date.
}
\examples{
# locate local OS downloads folder
downloads_dir <- vary::get_downloads_folder()
# write "Iris Report.xlsx" to downloads folder for example
openxlsx::write.xlsx(
  data.frame(iris),
  paste0(downloads_dir, "Iris Report.xlsx")
)

# search downloads for Excel files with "report" included in the title
files_matching(downloads_dir, "xlsx", name_contains = "report")

# search the R manual folder for pdf manuals (html also available)
r_manual_dir <- paste0(R.home("doc"), .Platform$file.sep, "manual")

pdf_manuals <- files_matching(r_manual_dir, "pdf")
pdf_manuals

# last modified (in this case, by the R installer)
latest_match <- pdf_manuals$path[1]
latest_match
}
