% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_desktop_folder.R
\name{get_desktop_folder}
\alias{get_desktop_folder}
\title{Get desktop folder}
\usage{
get_desktop_folder()
}
\value{
A character file path.
}
\description{
Return the path to the desktop directory regardless of operating system.
}
\examples{
get_desktop_folder()
}
