% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_downloads_folder.R
\name{get_downloads_folder}
\alias{get_downloads_folder}
\title{Get downloads folder}
\usage{
get_downloads_folder()
}
\value{
A character file path.
}
\description{
Return the path to the downloads directory regardless of operating system.
}
\examples{
get_downloads_folder()
}
