% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Negate value matching}
\usage{
abc \%notin\% xyz
}
\arguments{
\item{abc}{Vector or NULL: the values to be matched.}

\item{xyz}{Vector or NULL: the values to be matched against.}
}
\value{
A vector of the same length as x.
}
\description{
An intuitive counterpart to \code{\%in\%} which returns a logical
vector indicating if the left operand is \emph{not in} the right operand.
}
\examples{
1:10 \%notin\% c(1,3,5,9)
}
\seealso{
\code{\link[base:match]{base::match()}} \code{\link[base:funprog]{base::Negate()}}
}
