% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulchr-package.R
\docType{package}
\name{ulchr-package}
\alias{ulchr}
\alias{ulchr-package}
\title{ulchr: Uncategorized Personal Repo}
\description{
A collection of all the user-defined functions I commonly use, including both the well-documented functions from my other repos and any work-in-progress functions I'm currently developing.
}
\seealso{
Structure Data

\itemize{
\item \code{\link[=which_rows]{which_rows()}}
}

\itemize{
\item \code{\link[=fuzzy_match]{fuzzy_match()}}
}

\itemize{
\item \code{\link[=fuzzy_rename]{fuzzy_rename()}}
}

\itemize{
\item \code{\link[=flatten_pages]{flatten_pages()}}
}

\itemize{
\item \code{\link[=page_search]{page_search()}}
}

\itemize{
\item \code{\link[=drop_na_rows]{drop_na_rows()}}
}

\itemize{
\item \code{\link[=drop_na_cols]{drop_na_cols()}}
}

\itemize{
\item \code{\link[=cbind_ignore_row]{cbind_ignore_row()}}
}

Locate Files / Load Data

\itemize{
\item \code{\link[=get_downloads_folder]{get_downloads_folder()}}
}

\itemize{
\item \code{\link[=get_desktop_folder]{get_desktop_folder()}}
}

\itemize{
\item \code{\link[=files_matching]{files_matching()}}
}

\itemize{
\item \code{\link[=read_url_zip]{read_url_zip()}}
}

\itemize{
\item \code{\link[=upload_to_gsheets]{upload_to_gsheets()}}
}

Styling

\itemize{
\item \code{\link[=theme_ipsum_dot]{theme_ipsum_dot()}}
}

Other

\itemize{
\item \code{\link{\%notin\%}}
}

\itemize{
\item \code{\link[=view_DT]{view_DT()}}
}
}
\author{
\strong{Maintainer}: Chris Ulch \email{ulchchrisj@gmail.com}

}
\keyword{internal}
