% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_to_gsheets.R
\name{upload_to_gsheets}
\alias{upload_to_gsheets}
\title{Upload the latest downloaded file to Google Sheets}
\usage{
upload_to_gsheets(
  latest = "tsv",
  name_contains = "",
  sheetname = "Export",
  tabname = "Export"
)
}
\arguments{
\item{latest}{A string specifying a file type (csv or tsv).}

\item{name_contains}{A string used to filter for a specific file name.}

\item{sheetname}{The spreadsheet name.}

\item{tabname}{The tab name.}
}
\value{
The input ss, as an instance of googlesheets4::sheets_id
}
\description{
Uploads the latest file in the downloads folder of the specified type. A
string can be supplied to `name_contains` for a particular file.
}
\details{
NOTE: Requires remotes::install_github("ulchc/vary")
}
\examples{
upload_to_gsheets(latest = "tsv")
}
