% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_DT.R
\name{view_DT}
\alias{view_DT}
\title{Quickly view data with a DT::datatable}
\usage{
view_DT(
  df,
  randomize_index = FALSE,
  line_height = "50\%",
  max_cells = 1e+06,
  verbose = FALSE
)
}
\arguments{
\item{df}{A data frame or tibble.}
}
\value{
A DT::datatable of length nrow(df) - 1.
}
\description{
View your data while you code without switching to a tab created by View().
}
\seealso{
\code{\link[DT:datatable]{DT::datatable()}}
}
