% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulchr-package.R
\docType{package}
\name{ulchr-package}
\alias{ulchr}
\alias{ulchr-package}
\title{ulchr: Uncategorized Personal Repo}
\description{
A collection of all the user-defined functions I commonly use, including both the well-documented functions from my other repos and any work-in-progress functions I'm currently developing.
}
\seealso{
Structure Data

* [which_rows()]

* [fuzzy_match()]

* [fuzzy_rename()]

* [flatten_pages()]

* [page_search()]

* [drop_na_rows()]

* [drop_na_cols()]

Locate Files / Load Data

* [get_downloads_folder()]

* [get_desktop_folder()]

* [files_matching()]

* [read_url_zip()]

* [upload_to_gsheets()]

Styling

* [theme_ipsum_dot()]

Other

* [`%notin%`]
}
\author{
\strong{Maintainer}: Chris Ulch \email{ulchchrisj@gmail.com}

}
\keyword{internal}
